toolkit/content/tests/chrome/test_custom_element_base.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Custom Element Base Class Tests"
  onload="runTests();"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;"/>

  <simpleelement id="simple"/>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  SimpleTest.waitForExplicitFinish();

  function runTests() {
    ok(MozXULElement, "MozXULElement defined on the window");
    testParseXULToFragment();
    testCustomInterface();
    SimpleTest.finish();
  }

  function testParseXULToFragment() {
    ok(MozXULElement.parseXULToFragment, "parseXULToFragment helper exists");

    let frag = MozXULElement.parseXULToFragment(`<deck id='foo' />`);
    ok(frag instanceof DocumentFragment);

    document.documentElement.appendChild(frag);

    let deck = document.documentElement.lastChild;
    ok(deck instanceof MozXULElement, "instance of MozXULElement");
    ok(deck instanceof XULElement, "instance of XULElement");
    is(deck.id, "foo", "attribute set");
    is(deck.selectedIndex, "0", "Custom Element is property attached");
    deck.remove();
  }

  function testCustomInterface() {
    class SimpleElement extends MozXULElement {
      get disabled() {
        return false;
      }

      set disabled(val) {
      }

      get tabIndex() {
        return 0;
      }

      set tabIndex(val) {
      }
    }

    customElements.define("simpleelement", SimpleElement);
    MozXULElement.implementCustomInterface(SimpleElement, [Ci.nsIDOMXULControlElement]);

    is(document.documentElement.getCustomInterfaceCallback, undefined,
       "No getCustomInterfaceCallback on non-custom element");
    is(typeof document.getElementById("simple").getCustomInterfaceCallback, "function",
       "getCustomInterfaceCallback available on custom element when set");
    try {
      document.documentElement.QueryInterface(Ci.nsIDOMXULControlElement)
      ok(false, "Non-custom element implements custom interface");
    } catch (ex) {
      ok(true, "Non-custom element implements custom interface");
    }
    ok(document.getElementById("simple").QueryInterface(Ci.nsIDOMXULControlElement),
       "Implements custom interface");
  }
  ]]>
  </script>
</window>
